requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
PyYAML>=6.0.1